+++ src/mission/hs/+mag/+hs/+meta/InputSource.m
@@ -1,0 +1,10 @@
+classdef InputSource
+% INPUTSOURCE Type of input source for HelioSwarm data.
+
+    enumeration
+        % UART IMAP UART input source.
+        UART
+        % IDPU HelioSwarm iDPU input source
+        iDPU
+    end
+end